# ARROW-1342: [Python] Support strided ndarrays in pandas conversion from nested lists