@wesm 2017-08-09
The performance doesn't appear to be meaningfully different:

setup code

```
from collections import OrderedDict
import pyarrow as pa
import pandas as pd
import numpy as np
include_index = False

arrays = OrderedDict()
fields = []

fields.append(pa.field('int64', pa.list_(pa.int64())))
arrays['int64'] = [
    np.array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9]),
    np.array([0, 1, 2, 3, 4]),
    None,
    np.array([0])
]
fields.append(pa.field('double', pa.list_(pa.float64())))
arrays['double'] = [
    np.array([0., 1., 2., 3., 4., 5., 6., 7., 8., 9.]),
    np.array([0., 1., 2., 3., 4.]),
    None,
    np.array([0.])
]
fields.append(pa.field('str_list', pa.list_(pa.string())))
arrays['str_list'] = [
    np.array([u"1", u"ä"]),
    None,
    np.array([u"1"]),
    np.array([u"1", u"2", u"3"])
]

if include_index:
    fields.append(pa.field('__index_level_0__', pa.int64()))
df = pd.DataFrame(arrays)

df = pd.concat([df] * 10000, ignore_index=True)
```

before:

```
In [12]: timeit pa.Array.from_pandas(df['int64'])
13.6 ms ± 492 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

after:

```
In [2]: timeit pa.Array.from_pandas(df['int64'])
13.1 ms ± 729 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```