@wesm 2017-08-09
Updated the benchmarks, was accidentally picking up a debug library