@wesm 2017-08-09
Fixing the build, then going to merge this and cut an RC